SimpleProject/requirements.txt
fastapi==0.100.0
uvicorn==0.23.2
python-dotenv==1.0.0
pydantic==2.3.0
